Added ContactMessage functionalities.

+++ src/main/java/com/ColPlat/Backend/repository/UserRepository.java
@@ -4,10 +4,11 @@ import com.ColPlat.Backend.model.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, BigInteger> {
     Optional<User> findByEmail(String email);
     User findUserByEmail(String email);
 }
